Profil sayfası yapıldı sayfalar yapım aşamasında

+++ app/Http/Controllers/Back/CategoryController.php
@@ -33,35 +33,45 @@ class CategoryController extends Controller
 
     public function delete(Request $request)
     {
-
-        $category = Category::query()->find($request->id);
-
-        $article = Article::query()->where('category_id', '=', $request->id)->get();
-        if (count($article) > 0) {
-            flash()->preset('error', ['value' => $category->name . " kategorisini silerken", 'extra' => 'Öncelikle kategoriye ait makaleleri silmeniz gerekmektedir. ']);
+        if($request->id == 1){
+            flash()->error('_system Kategorisini silemezsin.');
             return redirect()->back();
-        } else {
+        }else {
+
             $category = Category::query()->find($request->id);
 
-            if (is_null($category)) {
-                flash()->error('Böyle bir kategori sistemde bulunmamaktadır.');
+            $article = Article::query()->where('category_id', '=', $request->id)->get();
+            if (count($article) > 0) {
+                flash()->preset('error', ['value' => $category->name . " kategorisini silerken", 'extra' => 'Öncelikle kategoriye ait makaleleri silmeniz gerekmektedir. ']);
+                return redirect()->back();
             } else {
-                $deleted = $category->delete();
-                if ($deleted) {
-                    flash()->success('Kategori başarıyla silindi.');
+                $category = Category::query()->find($request->id);
+
+                if (is_null($category)) {
+                    flash()->preset('not_found');
                 } else {
-                    flash()->error('Kategori silinirken hata oluştu.');
+                    $deleted = $category->delete();
+                    if ($deleted) {
+                        flash()->preset('deleted', ['value' => 'Kategori']);
+                    } else {
+                        flash()->preset('error', ['value' => 'Kategori silinirken']);
+                    }
                 }
+
+                return redirect()->back();
             }
 
-            return redirect()->back();
         }
-
-
     }
 
     public function forcedelete(Request $request)
     {
+        if($request->id == 1){
+            flash()->error('_system Kategorisini silemezsin.');
+            return redirect()->back();
+        }else{
+
+
         $action = $request->input('action');
         $category = Category::find($request->id);
 
@@ -92,30 +102,37 @@ class CategoryController extends Controller
         }
 
         return redirect()->back();
+        }
     }
 
     public function update(Request $request)
     {
-        $isSlug = Category::whereSlug(Str::slug($request->slug))->whereNotIn('id', [$request->id])->first();
-        $isName = Category::whereName($request->category)->whereNotIn('id', [$request->id])->first();
-        if ($isSlug or $isName) {
-            flash()->error($request->category . 'Aynı isimde/linkte kategori sistemde bulunmaktadır.');
-            return redirect()->route('admin.category.index');
+        if($request->id == 1){
+            flash()->error('_system kategorisini güncelleyemezsin!');
+            return redirect()->back();
+        }else{
+            $isSlug = Category::whereSlug(Str::slug($request->slug))->whereNotIn('id', [$request->id])->first();
+            $isName = Category::whereName($request->category)->whereNotIn('id', [$request->id])->first();
+            if ($isSlug or $isName) {
+                flash()->error($request->category . 'Aynı isimde/linkte kategori sistemde bulunmaktadır.');
+                return redirect()->route('admin.category.index');
+            }
+
+            $category = Category::query()->find($request->id);
+            $category->name = $request->category;
+            $category->slug = STR::slug($request->slug);
+            $category->save();
+            flash()->success('Kategori güncelleme İşlemi Başarılı');
+            return redirect()->back();
         }
 
-        $category = Category::query()->find($request->id);
-        $category->name = $request->category;
-        $category->slug = STR::slug($request->slug);
-        $category->save();
-        flash()->success('Kategori güncelleme İşlemi Başarılı');
-        return redirect()->back();
     }
 
     public function create(Request $request)
     {
         $isExist = Category::whereSlug(Str::slug($request->category))->first();
         if ($isExist) {
-            toastr()->error('Aynı isimde kategori sistemde bulunmaktadır.');
+            flash()->error('Aynı isimde kategori sistemde bulunmaktadır.');
             return redirect()->route('admin.category.index');
         }
 
@@ -123,7 +140,7 @@ class CategoryController extends Controller
         $category->name = $request->category;
         $category->slug = STR::slug($request->category);
         $category->save();
-        toastr()->success('Kategori ekleme İşlemi Başarılı');
+        flash()->success('Kategori ekleme İşlemi Başarılı');
         return redirect()->route('admin.category.index');
     }
 }
